requirements.txt
pybit~=2.4.0
setuptools~=67.4.0
Cython~=0.29.33
Flask~=2.1.2
gunicorn~=20.1.0
python-dotenv~=1.0.0
python-telegram-bot~=20.2